macos_controller/pyproject.toml
[project]
name = "mcp_remote_macos_use"
version = "0.1.0"
description = "A MCP server for remote MacOS control"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.4.1",
    "python-dotenv>=1.0.1",
    "pillow>=10.0.0",
    "AsyncVNC>=0.2.0",
    "anthropic>=0.49.0",
    "paramiko>=3.5.1",
    "livekit>=1.0.5",
    "aiohttp>=3.8.1",
    "websockets>=10.0",
    "aiortc>=1.3.2",
    "livekit-api>=1.0.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = ["pyright>=1.1.389", "pytest>=7.4.0", "pytest-asyncio>=0.21.1"]

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_remote_macos_use"]

[project.scripts]
mcp_remote_macos_use = "mcp_remote_macos_use:main" 